@@ -36,11 +36,8 @@ def create_app(testing=False):
     register_blueprints(app)
 
     # Set up CORS
-    CORS(app, resources={
-        r"/api/*": {
-            "origins": ["http://10.0.1.3:8080", "http://localhost:8083"]
-        }
-    })
+    CORS(app)
+    # CORS(app, resources={r"/api/*": {"origins": "*"}})  # Możesz ograniczyć do konkretnych źródeł zamiast używać "*"
 
     # Define routes
     @app.route('/api')
