@@ -1,25 +1,6 @@
-// frontend/src/main.js
+import './assets/main.css'
+
 import { createApp } from 'vue'
 import App from './App.vue'
-import * as Sentry from "@sentry/vue";
-
-Sentry.init({
-  dsn: "https://1a5e64e0189716a59022f518bcf35c69@o4506447864463360.ingest.sentry.io/4506453239398400",
-  integrations: [
-    new Sentry.BrowserTracing({
-      // Set 'tracePropagationTargets' to control for which URLs distributed tracing should be enabled
-      tracePropagationTargets: ["localhost", /^https:\/\/yourserver\.io\/api/],
-    }),
-    new Sentry.Replay({
-      maskAllText: false,
-      blockAllMedia: false,
-    }),
-  ],
-  // Performance Monitoring
-  tracesSampleRate: 1.0, //  Capture 100% of the transactions
-  // Session Replay
-  replaysSessionSampleRate: 1.0, // This sets the sample rate at 10%. You may want to change it to 100% while in development and then sample at a lower rate in production.
-  replaysOnErrorSampleRate: 1.0, // If you're not already sampling the entire session, change the sample rate to 100% when sampling sessions where errors occur.
-});
 
 createApp(App).mount('#app')
